Cleaned/standardized TempBlitz times/locations
Start and end times standardized for the August Blitz values. Not sure if you wanted AM/PM at the end or if you wanted it in 12hr/24hr, or if it's even helpful at all, but regardless it was good practice.

Also cleaned the user inputted pole/ logger values and standardized with the HoboLocations_PoleLocations.csv. Noted all changes and displayed a grouped version of the data at the end of the Jupyter Notebook to help visualize.